Code reorg to try and fix the go generate ordering problem in CI.
Also because types.go got too big.
